# swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: BookLibrary
  description: |
    books available for loan
  license:
    name: MIT
servers:
- url: http://localhost:3000
paths:
...
  /books/{bookId}:
    get:
          summary: Info on specific book
          operationId: getBookById
          parameters: 
          - name: bookId
            in: query
            description: Book ID
            required: true
            schema:
              type: string
              format: int64
          tags:
          - books
          responses:
            '200':
              description: Show one book
              content:
                application/json:    
                  schema:
                    $ref: "#/components/schemas/ResponseBooks"
            default:
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/Error"
    patch:
          summary: Updating specific book
          operationId: updateBook
          parameters:
          - name: bookId
            in: path
            description: Book ID
            required: true
            schema:
              type: string
              format: int64
          tags:
          - books
          requestBody:
            description: the book to create
            required: true
            content:
              application/json:    
                schema:
                  $ref: "#/components/schemas/Book"
          responses:
            '201':
              description: req.param upserted
              content:
                application/json:    
                  schema:
                    $ref: "#/components/schemas/ResponseBooks"
            '200':
              description: OK
              content:
                application/json:    
                  schema:
                    $ref: "#/components/schemas/ResponseBooks" 
            '204':
              description: No content
              content:
                application/json:    
                  schema:
                    $ref: "#/components/schemas/ResponseBooks"              
            default:
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/Error"
    delete:
          summary: Delete specific book
          operationId: deleteBook
          parameters:
          - name: bookId
            in: path
            description: Book ID
            required: true
            schema:
              type: string
              format: int64 
          tags:
          - books
          responses:
            '200':
              description: deleted
              content:
                application/json:    
                  schema:
                    $ref: "#/components/schemas/ResponseBooks"
            '204':
                description: Cant find id
                content:
                  application/json:    
                    schema:
                      $ref: "#/components/schemas/Error"                 
            default:
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Book: # A single book
      type: object
      # allOf:
      # - $ref: "#/components/schemas/Book"
      properties:
          ISBN:
            type: string
          Title:
           type: string
          Author:
            type: string
          Price:
           type: integer
          SellerEmail:
           type: string
          Used:
           type: boolean
          Location:
           type: object
           properties:
            City:
              type: string
            Street:
              type: string
# Add properties for your books here   
    ResponseBook: #Single book response including _id and __v
      type: object
      allOf:
      - $ref: "#/components/schemas/Book"
      - properties:
          _id:
            type: string
            format: uuid
          __v:
            type: integer
            format: uint32
    ResponseBooks: #list of books
      type: array
      items:
        $ref: "#/components/schemas/ResponseBook"
    Error:
      required:
        - error
      properties:
        error:
          type: object
          properties:
            expose:
              type: boolean
            statusCode:
              type: integer
              format: uint32
            status:
              type: integer
              format: uint32
            body:
              title: body as escaped json
              type: string
            type:
              title: fault desciption
              type: string
              $ref: "#/components/schemas/Error"